requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Google Gemini API
google-generativeai==0.3.2

# Background Jobs
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Frontend
jinja2==3.1.3